backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==1.4.50
psycopg2-binary==2.9.9
asyncpg==0.29.0
stripe==7.8.0
python-dotenv==1.0.0
email-validator==2.1.0
